removed submitscoremodal and renamed to submitscorebutton
